Application status added to flash loader menu
This will show information about the application
including its size and start address.
It will also show the application valid pattern.
The application is valid if it is 5A5A5A5A.
If no application has been loaded it is FFFFFFFF
The status of the prompt GIO pin is also shown here.